Move KeyboardAction and KeyboardData to structures folder
