+++ 8vim/src/main/java/inc/flide/vim8/keyboardHelpers/InputMethodServiceHelper.java
@@ -10,6 +10,8 @@ import java.util.Map;
 import inc.flide.vim8.R;
 import inc.flide.vim8.preferences.SharedPreferenceHelper;
 import inc.flide.vim8.structures.FingerPosition;
+import inc.flide.vim8.structures.KeyboardAction;
+import inc.flide.vim8.structures.KeyboardData;
 
 public class InputMethodServiceHelper {
