+++ 8vim/src/main/java/inc/flide/vim8/structures/KeyboardData.java
@@ -1,10 +1,11 @@
-package inc.flide.vim8.keyboardHelpers;
+package inc.flide.vim8.structures;
 
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 import inc.flide.vim8.structures.FingerPosition;
+import inc.flide.vim8.structures.KeyboardAction;
 
 public class KeyboardData {
     private Map<List<FingerPosition>, KeyboardAction> actionMap;
